+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Breaking Change
 - elastic_class.ElasticSearch.\_\_init\_\_:  Replaced attribute updates with call to update_status().
 
 ### Added
-- elastic_class.ElasticSearchDump.update_status:  Update class attributes by querying Elasticsearch.
+- elastic_class.ElasticSearchDump.update_dump_status:  Update class attributes by querying Elasticsearch.
 - elastic_class.ElasticSearchStatus:  Added class and associated methods.
 - elastic_class.ElasticSearch.update_status:  Update class attributes by querying Elasticsearch.
 - elastic_class.create_snapshot:  Runs a dump of a repository.